Infographic Redesign
Added a scroll feature to the infographic guide so that all of the fruit sections can fit on the screen as well as minor cosmetic and picture changes. Added a new section for strawberries on the guide. Added some error handling for outputting text to the screen after the computations.